Refine notion of children: there are now immediate and non-immediate children
An immediate child is the direct child of a node. For symmetry, we also distinguish
immediate parents, but that is not as important.
With this, we can change the tag invariant to only talk about immediate children
of owned tags. This might make it possible to recover unique ownership
whenever a tag is written.