Fix logical_state invariant for unique tags.
The precondition for tk_unq tags was too strong,
since it required non-protected tags to not be frozen.
This requirement is not showable when we want to do a read from such a tag,
since in that case we only know the tag was readable.
By moving this precondition around a bit we are able to learn enough during
a read to conclude that we read the right values.